# WIP: Optimize for big repository on repo home page to make it load less than 1s and /main less than 8s