--- a/modules/cache/lastcommit/init.go
+++ b/modules/cache/lastcommit/init.go
@@ -11,7 +11,7 @@ import (
 )
 
 var (
-	// LastCommitCache defines gloablly last commit cache object
+	// LastCommitCache defines globally last commit cache object
 	LastCommitCache git.LastCommitCache
 )