--- a/modules/cache/lastcommit/init.go
+++ b/modules/cache/lastcommit/init.go
@@ -5,7 +5,10 @@
 package lastcommit
 
 import (
+	"time"
+
 	"code.gitea.io/git"
+	"code.gitea.io/gitea/modules/cache"
 	"code.gitea.io/gitea/modules/log"
 	"code.gitea.io/gitea/modules/setting"
 )
@@ -17,12 +20,27 @@ var (
 
 // NewContext init
 func NewContext() error {
+	if cache.Cache != nil && setting.Git.LastCommitCache.UseDefaultCache {
+		LastCommitCache = &lastCommitCache{
+			mc:      cache.Cache,
+			timeout: int64(setting.CacheService.TTL / time.Second),
+		}
+		return nil
+	}
+
 	var err error
 	switch setting.Git.LastCommitCache.Type {
 	case "memory":
 		LastCommitCache = &MemoryCache{}
 	case "boltdb":
-		LastCommitCache, err = NewBoltDBCache(setting.Git.LastCommitCache.DataPath)
+		LastCommitCache, err = NewBoltDBCache(setting.Git.LastCommitCache.ConnStr)
+	case "redis":
+		addrs, pass, dbIdx, err := parseConnStr(setting.Git.LastCommitCache.ConnStr)
+		if err != nil {
+			return err
+		}
+
+		LastCommitCache, err = NewRedisCache(addrs, pass, dbIdx)
 	}
 	if err == nil {
 		log.Info("Last Commit Cache %s Enabled", setting.Git.LastCommitCache.Type)