… docs/content/doc/advanced/config-cheat-sheet.en-us.md
… docs/content/doc/advanced/config-cheat-sheet.zh-cn.md
… modules/setting/cache.go
… modules/setting/git.go
… modules/setting/setting.go
… modules/templates/helper.go
… routers/init.go
… routers/repo/editor.go
… routers/repo/issue.go
… routers/repo/view.go
… routers/repo/wiki.go
… templates/repo/view_list.tmpl
… vendor/code.gitea.io/git/cache.go
… vendor/code.gitea.io/git/commit.go
… vendor/code.gitea.io/git/commit_info.go
… vendor/code.gitea.io/git/tree.go
… vendor/code.gitea.io/git/tree_blob.go
… vendor/code.gitea.io/git/tree_entry.go
… vendor/go.etcd.io/bbolt/LICENSE
… vendor/go.etcd.io/bbolt/bolt_386.go
… vendor/go.etcd.io/bbolt/bolt_amd64.go
… vendor/go.etcd.io/bbolt/bolt_arm.go
… vendor/go.etcd.io/bbolt/bolt_arm64.go
… vendor/go.etcd.io/bbolt/bolt_linux.go
… vendor/go.etcd.io/bbolt/bolt_mips64x.go
… vendor/go.etcd.io/bbolt/bolt_mipsx.go
… vendor/go.etcd.io/bbolt/bolt_openbsd.go
… vendor/go.etcd.io/bbolt/bolt_ppc.go
… vendor/go.etcd.io/bbolt/bolt_ppc64.go
… vendor/go.etcd.io/bbolt/bolt_ppc64le.go