I tested it locally. The recent linux repo home page loads time from 8s to 2s if enabled memory last commit cache, 8s to 2.8s if enabled boltdb last commit cache.

**update**: now it only 1.9s even boltdb on my MacOS since we only get simple commit infomation.

**update(2019-02-22)**: added ls tree cache and files limit to 1000, now aport will /main directory will load less than 8s on my Macbook pro. I will add a log cache and need more work to clean up the codes.

blocked by https://github.com/go-gitea/git/pull/145