@lafriks 2019-02-12
Why not just reuse already existing caching functionality like commit counts are cached?