@lunny 2019-02-12
I think that cache infrastructures are too complicated. and I think the commit counting cache should also be used a separated cache service to do that.