@clandmeter 2019-02-12
Thanks a lot @lunny for trying to improve this usability concern.

I can confirm that using boltdb improves performance the most for me.
```
memory:
	inital load 120 seconds
	second load 60-70 seconds

boltdb:
	second load 40-50 seconds

none:
	inital load 120 seconds
	second load 130 seconds
```

url to try is http://51.15.253.24:3000/clandmeter/aports/src/branch/master/main

What really would make a difference on repo layouts like this would be a pager with a commit list limit set in app.ini. Or if that's not possible a hard limit like github does.