@lunny 2019-02-12
@clandmeter have you enabled `[cache]` on your app.ini. That will also help on commit counting. We have more improvements work to do. Maybe we should learn how github did that. limit 1000 files and hide all the last commit message?