@clandmeter 2019-02-12
@lunny i enabled cache (redis) as this was a vanilla setup. the load is now around 30-40 seconds.

Yes a hard limit could be an option.
If the limit amount could be a setting users would be able to fine tune it based on system spec.

Github does the same: https://github.com/alpinelinux/aports/tree/master/main

This loads in a few seconds 1-3. I don't think gitea can do that even with a 1k limit. The amount of objects is now around 2k so i would assume at best it would load in 15 seconds.

Even with caching I see a lot of git commands (git rev-list -1 xxx) beeing executed in the background. I guess the cache doesn't cover all of the git lookups.