@lunny 2019-02-13
@clandmeter yes. I need send several PRs for all the optimizations since not all the git command be cached. I will send another PR to limit the display files number according user setting.