@clandmeter 2019-02-13
@lunny I wonder why you choose to use boltdb but did not add a Redis option? Redis is already used for caching and would provide much better scaling when your instance would grow?