@lafriks 2019-02-13
@lunny imho it would be better to implement type that uses internally existing caching methods. This way we would support also redis etc