@lunny 2019-02-13
@clandmeter It's easy to add redis.
@lafriks Will consider that.