@lunny 2019-02-16
@lafriks added old cache configuration support.