@lunny 2019-02-16
@clandmeter added redis support and reduced half of git commands when list last commit information.