@clandmeter 2019-02-16
@lunny nice thx.
What is the config option to enable redis for commit cache?
Also applying this PR on master will segfault when using boltdb.

```
[Macaron] 2019-02-16 23:29:38: Started GET /clandmeter/aports/src/branch/master/main for 172.16.254.10
panic: runtime error: invalid memory address or nil pointer dereference
[signal SIGSEGV: segmentation violation code=0x1 addr=0x28 pc=0x55cf09b37a95]

goroutine 1043 [running]:
code.gitea.io/gitea/modules/cache/lastcommit.(*BoltDBCache).Get(0x0, 0xc001a12540, 0x31, 0xc001a23b90, 0x28, 0xc0036648d0, 0x9, 0xc000513708, 0x55cf08d95855, 0xc000a35c80)
        /home/clandmeter/go/src/code.gitea.io/gitea/modules/cache/lastcommit/boltdb.go:57 +0xe5
code.gitea.io/gitea/vendor/code.gitea.io/git.targetedSearch(0xc000c86280, 0xc0024263c0, 0x55cf0a7f7860, 0x0)
        /home/clandmeter/go/src/code.gitea.io/gitea/vendor/code.gitea.io/git/commit_info.go:105 +0x5aa
created by code.gitea.io/gitea/vendor/code.gitea.io/git.getCommitsInfo
        /home/clandmeter/go/src/code.gitea.io/gitea/vendor/code.gitea.io/git/commit_info.go:247 +0x29c

```