@clandmeter 2019-02-17
@lunny could you update the ini sample on how to use the new options? I am unable to make redis work. I can only make memory work.