@lunny 2019-02-22
@clandmeter I send another commit today and I think that will make aport `/main` loading less than 8s(max 1000 files). And I will add another cache soon.