@clandmeter 2019-02-22
@lunny im not able to make this work anymore.
What is the correct way to add redis caching?
I used:

```
; Last commit cache
[git.last_commit_cache]
USE_DEFAULT_CACHE = true

[cache]
ADAPTER = redis
HOST = network=tcp,addr=:6379,db=0,pool_size=100,idle_timeout=180
```

which results in

```
[Macaron] 2019-02-22 17:02:14: Started GET /alpine/aports for 172.16.254.10
panic: interface conversion: interface {} is string, not *git.Commit

goroutine 1049 [running]:
code.gitea.io/gitea/modules/cache/lastcommit.(*lastCommitCache).Get(0xc00031a9e0, 0xc001989b00, 0x2d, 0xc001a2b5c0, 0x28, 0xc003a3f4a0, 0x9, 0x0, 0x0, 0x0)
        /home/clandmeter/go/src/code.gitea.io/gitea/modules/cache/lastcommit/cache.go:23 +0x103
code.gitea.io/gitea/vendor/code.gitea.io/git.targetedSearch(0xc00003b630, 0xc003a55620, 0x563ca2460060, 0xc00031a9e0)
        /home/clandmeter/go/src/code.gitea.io/gitea/vendor/code.gitea.io/git/commit_info.go:105 +0x5aa
created by code.gitea.io/gitea/vendor/code.gitea.io/git.getCommitsInfo
        /home/clandmeter/go/src/code.gitea.io/gitea/vendor/code.gitea.io/git/commit_info.go:247 +0x29c
```