@typeless 2019-02-23
I am wondering if [gitbase](https://github.com/src-d/gitbase) could be helpful here.
That way, we can use xorm on top of it and freeride on xorm's caching mechanism or other existing caching library for databases. 

I haven't carefully looked into gitbase, so this might be a naive idea. 😆