@lunny 2019-02-23
@typeless I don't know if gitbase is useful and related. For this PR, it should resolve all big repo home page issue. On my test, aports will only spent 800ms after throughout cached. But now there are some confusing issues I will resolve.