@clandmeter 2019-02-23
@lunny my observations with current code:

Using cache in memory gets page load to 500ms, impressive :)

Using boltdb is not working when using the same db for both git caches, gitea just wont start is this actually supported, if not maybe put that in the docs. If i use separate dbs than the result is as with redis.

Using Redis i dont see any improvements, page loads now in 15 seconds, before was 30 seconds but only 1000 items are loaded.

I would be nice if we could configure the 1000 limit or even disable it when caching is working properly.

config:

```
[cache]
ADAPTER = redis
HOST = network=tcp,addr=:6379,db=0,pool_size=100,idle_timeout=180

[cache.last_commit]
TYPE = redis
ENABLE_MIN_COMMIT_COUNT = 0
CONN_STR = addrs=127.0.0.1 password= db=0

[cache.ls_tree]
TYPE = redis
ENABLE_MIN_COMMIT_COUNT = 0
CONN_STR = addrs=127.0.0.1 password= db=0
```