@lunny 2019-02-23
@clandmeter Glad to hear that memory cache could be used. I have known that `boltdb` has some problems and I'm investigating that to fix it. I haven't tested `redis` so I don't know why it's slow. Of cousre, the limitation of show files will be added as a config item. I will continue the work. Thanks for your testing.