@clandmeter 2019-02-23
@lunny yes after retesting boltdb it seems broken.

Regarding redis, it looks like it doesn't cache anything. Compared to memory caching you can see the second run will do none or almost no git lookups anymore. With redis this keeps happening making it much slower. Funny thing is is does check for redis availability because if i turn it of it will return error 500.

I have some small remarks for your reference.

To my understanding a redis lookup should be much faster than a git lookup. So if you enable global cache with redis why not automatically enable it for the other cache options?

Also what does the following mean:

> repos which have less than commits number will enable this cache, 0 will always cache if TYPE is not empty

This reads as repo with more than 1000 commits (default setting) have caching disabled? 
The amount of commits in a tree has nothing to do with this caching option, or should i read this differently?

Thanks for looking into it, and ill keep following this PR.